🔨 Refactor: DiaryContent jotai로 전환
