+++ src/pages/Diary/components/DiaryContent/DiaryContentEditButton.tsx
@@ -1,8 +1,17 @@
-import useDiaryContentContext from '../../hooks/DiaryContent/useDiaryContentContext';
+// import useDiaryContentContext from '../../hooks/DiaryContent/useDiaryContentContext';
+import { useAtomValue } from 'jotai';
 import { Button } from '~/components/common';
+import useDiaryContents from '~/pages/Diary/hooks/DiaryContent/useDiaryContents';
+import {
+  editDiaryAtom,
+  editableAtom,
+  loadingAtom,
+  originDiaryAtom,
+} from '~/stores/diaryContentAtoms';
 const DiaryContentEditButton = () => {
-  const { editable, methods, loading } = useDiaryContentContext();
-  const { handleEditCancel, handleSubmitForm } = methods;
+  const editable = useAtomValue(editableAtom);
+  const loading = useAtomValue(loadingAtom);
+  const { handleEditCancel, handleSubmitForm } = useDiaryContents();
 
   return (
     editable && (
