+++ src/pages/Diary/components/DiaryContent/DiaryContentHeader.tsx
@@ -1,14 +1,25 @@
+import { useAtomValue } from 'jotai';
 import { memo } from 'react';
 import { paths } from '~/router';
-import useDiaryContentContext from '../../hooks/DiaryContent/useDiaryContentContext';
+// import useDiaryContentContext from '../../hooks/DiaryContent/useDiaryContentContext';
 import { DiaryHeader } from '~/pages/Diary/components/DiaryCommon';
-import useDiaryContext from '~/pages/Diary/hooks/Diary/useDiaryContext';
+// import useDiaryContext from '~/pages/Diary/hooks/Diary/useDiaryContext';
+import useDiaryContents from '~/pages/Diary/hooks/DiaryContent/useDiaryContents';
+import { infoAtom } from '~/stores/diaryAtoms';
+import {
+  editDiaryAtom,
+  editableAtom,
+  originDiaryAtom,
+} from '~/stores/diaryContentAtoms';
 
 const DiaryContentHeader = memo(() => {
-  const { editable, diary, methods } = useDiaryContentContext();
-  const { info } = useDiaryContext();
+  const editable = useAtomValue(editableAtom);
+  const editDiary = useAtomValue(editDiaryAtom);
+  const originDiary = useAtomValue(originDiaryAtom);
+  const info = useAtomValue(infoAtom);
+  const diary = editable ? editDiary : originDiary;
   const { kakaoMapId, placeName } = diary;
-  const { handleEditMode, handleDeleteDiary } = methods;
+  const { handleEditMode, handleDeleteDiary } = useDiaryContents();
   const { DIARY } = paths;
   const prevLink = info ? `${DIARY.ROOT}/${kakaoMapId}` : `${DIARY.ROOT}`;
