+++ src/pages/Diary/components/DiaryContent/DiaryContentImgs.tsx
@@ -1,9 +1,11 @@
-import useDiaryContentContext from '../../hooks/DiaryContent/useDiaryContentContext';
+// import useDiaryContentContext from '../../hooks/DiaryContent/useDiaryContentContext';
+import { useAtomValue } from 'jotai';
 import DiaryImgsCarousel from './DiaryImgsCarousel';
 import DiaryImgsUpload from './DiaryImgsUpload';
+import { editableAtom } from '~/stores/diaryContentAtoms';
 
 const DiaryContentImgs = () => {
-  const { editable } = useDiaryContentContext();
+  const editable = useAtomValue(editableAtom);
 
   return <>{editable ? <DiaryImgsUpload /> : <DiaryImgsCarousel />}</>;
 };
